+++ Automation-server/BLE-CCS811.h
@@ -3,4 +3,4 @@
 
 
 bool setupCCS811();
-void pollCCS811(float temp, float hum);
+bool pollCCS811(float temp, float hum);